requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
dj-database-url==2.1.0
openai>=1.0.0
requests>=2.31.0
djangorestframework>=3.14.0